requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
black==24.1.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Django==5.0.1
django-cors-headers==4.3.1
django-ninja==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.4.0
flake8==7.0.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
mccabe==0.7.0
msgpack==1.0.8
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.18
psycopg-binary==3.1.18
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pyflakes==3.2.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
